Explicitly cast stat::st_size field to std::size_t

+++ GOESP/Memory.cpp
@@ -55,7 +55,7 @@ static std::pair<void*, std::size_t> getModuleInformation(const char* name) noex
 
         if (const auto fd = open(info->dlpi_name, O_RDONLY); fd >= 0) {
             if (struct stat st; fstat(fd, &st) == 0) {
-                if (const auto map = LinuxFileMap{ nullptr, st.st_size, PROT_READ, MAP_PRIVATE, fd, 0 }; map.isValid()) {
+                if (const auto map = LinuxFileMap{ nullptr, static_cast<std::size_t>(st.st_size), PROT_READ, MAP_PRIVATE, fd, 0 }; map.isValid()) {
                     const auto ehdr = (ElfW(Ehdr)*)map.get();
                     const auto shdrs = (ElfW(Shdr)*)(std::uintptr_t(ehdr) + ehdr->e_shoff);
                     const auto strTab = (const char*)(std::uintptr_t(ehdr) + shdrs[ehdr->e_shstrndx].sh_offset);
